intermediate version #1

+++ azure_c_shared_utility/socketio_mbed.c
@@ -331,7 +331,7 @@ int socketio_send(CONCRETE_IO_HANDLE socket_io, const void* buffer, size_t size,
             else
             {
                 int send_result = tcpsocketconnection_send(socket_io_instance->tcp_socket_connection, buffer, size);
-                if (send_result != (int)size)
+                if (send_result != size)
                 {
                     if (send_result < 0)
                     {
@@ -385,7 +385,7 @@ void socketio_dowork(CONCRETE_IO_HANDLE socket_io)
                 }
 
                 int send_result = tcpsocketconnection_send(socket_io_instance->tcp_socket_connection, (const char*)pending_socket_io->bytes, pending_socket_io->size);
-                if (send_result != (int)pending_socket_io->size)
+                if (send_result != pending_socket_io->size)
                 {
                     if (send_result < 0)
                     {
